Reorganize chatbot functions to 'chatbot_logic.py'
